"Redbol" compatibility changes for Rebol variant
The Ren-C branch of Rebol3 has tons of changes.  However, it seeks to
be versatile enough to be skinned with an emulation layer that can run
many R3-Alpha, Rebol2, and Red programs.

Programs may need adjustment before being able to run under emulation.
The idea of these adjustments is that the adjusted code can still run
under its initial medium with them in effect.  This makes those minor
adjustments.

1. The unpopular "++" and "--" constructs from R3-Alpha are deliberately
not emulated by default, in order to use the debugging-oriented feature
(`-- var` will dump the variable `var` and show its name, as well as be
invisible to the surrounding evaluations).  Uses ordinary increment
via + instead of the increment operator.

2. CHAR! is no longer a distinct datatype from ISSUE!.  The two are
unified so that #x can be a plain character literal.  Short issues are
compacted into single cells as their UTF-8 payloads.  But because of
this, functions like REJOIN which are cued by their first element will
now form an ISSUE! if what used to be a "CHAR!" is the first element.
AJOIN is a better compatibility function if a string was intended
from the join.