Use MOLD instead of plain PRINT for program output
Changed the completely invisible PRINT to PRINT MOLD...which still
doesn't help you with seeing whitespace at end of line.  But is vaguely
more calibrating than just seeing full empty space.

(Better would be to offer the option of transforming to the symbolic
tab space newline, which is suggested as a future option.)