Separate copy of Rebol code for Ren-C Version
The decade-old Rebol version of %whitespace.reb was a newbie effort in
the language, which only hints at the kinds of tricks a Rebol-based
solution to the problem might use.

This puts aside that version for historical reference, and starts a
new file which runs without an emulation layer under the Ren-C branch:

https://github.com/metaeducation/ren-c

The file is unchanged in this commit, so that the modifications
required to get it to run can be seen in isolation.