Make OPERATION use a more succinct format
Instead of making OPERATION a simple alias for MAKE OBJECT! where
each instance of the operation had to repeat the `command:` and
`description:` keys, this shifts to assuming that the first element
of the block is a TEXT! string...and the remaining elements are the
command definition.

This cuts down significantly on repetition.