Make OPERATION create the handler function
The initial step of moving the Rebol code for the opcode handlers
to be part of the operation definition would pass the already-made
function as a parameter to OPERATION.  However, the goal is to derive
the argument list to that function from the whitespace spec.  So the
spec of the function would be merged with the `space lf tab` sequence
information, hence OPERATION would be creating the action.

This moves one step toward that by putting the FUNC call inside of
OPERATION to do the generation.  To get a small benefit from this,
the body is augmented to say `return null` in case there is no explicit
return already in the function.

(It's likely desirable to not do the jumping via RETURN values at all,
but by something more obvious...like an implicit MACHINE parameter
that can have its instruction pointer queried and set.  Either way,
the RETURN NULL that operations were being forced to end with is
too redundant and should be avoided.)