Integrate parameter definitions into operation spec
Previously the function argument definition was distinct from the
specification in the operation that it took a Number or a Label as
a parameter.  This blends the two with a parameter definition syntax
that uses a BLOCK! with a SET-WORD! of the name, followed by the
WORD! of either `Number` or `Label`, e.g.

    push: operation [
        {Push the number onto the stack}
        space [value: Number]
    ][
        insert stack value
    ]