GitHub Actions workflow to test Rebol Whitespacers
While this repository is a collection of interpreters for multiple
languages, the collection was made as a survey to inform the
development of the Rebol whitespace interpreter that it was started
for.  This adds a test to support checking that both a historical
version and a modern version will run a test program.

It uses the Ren-C GitHub Action:

https://github.com/metaeducation/ren-c-action