Reorganize R3-Alpha/Ren-C interpreters from collection
A Rebol whitespace interpreter was published along with a
collection of other interpreters from the Internet.  That small
repository is now superseded by a larger and more formal project
to collect interpreters from various languages (along with getting
instructions for running them).

Since the Ren-C interpreter is now considered an interesting use
case for the language, it is undergoing development and needs a
repository of its own.  This goes ahead and does that, and the
original repository is shifting to a read-only archive.