Take filename on the command line
The original whitespace interpreter had hardcoded test data in it,
and never got to the point where it took command-line parameters.

This makes it possible to take either a .ws file or a new ".wsw"
format (which is just the words `space`, `tab`, and `lf` and can
allow comments with semicolons).