Don't print newlines on char/number output
In trying to actually make something that can run the test programs
without debug information in it, the fact that it was printing newlines
after every character and number was observed.

This changes to use WRITE-STDOUT instead of PRINT (which always puts
a newline if the output block does not completely vaporize).  e.g.
with PRINT:

    print [""] -> prints just a newline
    print [] -> outputs nothing