Filter out non-whitespace characters
The original tutorial program did not have any non-whitespace characters
in it.  But it is central to the whitespace language to be able to fit
programs "out of band" inside of other programs.

Amending the parse rules to skip over the characters would be difficult
as it would have to do so between every character scan in every rule.
It would also complicate decoding.  The easiest thing is just to remove
the characters from the loaded program.