Don't keep spec tags in parse rules
There was a facility for letting you put things like `<local>` and
`<static>` in the spec of the operation handlers, just like you could
do for a normal function.  This shouldn't appear in the parse rule
for the command.  However it was being included in the COLLECT for
that for some reason.

This fixes that, and since there were no examples to show it wasn't
working, this adds one by using `<local>` in the spec for the mark
location operation.