Change labels to use strings and not numeric conversion
The same rule was being used for labels as for numbers, as they were
both spaces and tabs ended by a line feed.  However, labels can be
arbitrarily long (e.g. they could be encodings of text strings in
binary).  So they would require bignums to store.

While Ren-C does have some BigNum code in the works, it's not ready.
But also unnecessary for this.  So this switches labels to just being
the text of the whitespace string.