Handle empty callstack by noticing NULL TAKE
For a time, TAKE of an empty block would cause an error.  It was
decided that returning NULL was the better option, and those wishing
to ensure it was not null were responsible for either writing an
ELSE clause or saying `non null take`.

This means it's cleaner to handle the empty callstack condition.