Add READ-CHARACTER-TO-LOCATION
The desktop I/O system that Ren-C inherited from R3-Alpha is bad at
handling piping (and not relevant to the current primary target, which
is Wasm in the browser).

But for the sake of passing the name.ws test, reading piped characters
has been added...hacked in hopefully well enough to make the test work.