Use GATHER/EMIT to parse operation specs, more ANY uses
The more recent iteration of the whitespacers is based on the idea
of OPERATION which has a spec like `[space tab [argname: Label]]`,
but can also be mixed with information that is transferred onto the
spec of the function behind the operation's implementation.

When the code to process that was first written UPARSE did not exist.
Now that it does, it's the obvious choice for processing the spec.

This transitions it to not only use UPARSE in the spec processing, but
to embrace things like ANY and the COLLECT features in rules that
are generated by the spec which--in turn--generate the decoded
instruction blocks like [push 10].