Make OPERATION bind the rule into itself
The instruction block being generated by OPERATION would look good
as `[push 304]`, but that PUSH wasn't bound to anything.  So if you
tried to DO the instruction it would error.

The VM went through some pains to try and figure out how to bind that
to the function in the operation.  Which doesn't make much sense
anymore, as now the operation can simply do the binding at the
time of creation.

That small tweak lets us get rid of some ugly hackery in the VM
to try and get the instruction block bound.