Initial modularization attempt, add %ws-common.reb
R3-Alpha had a very unfinished module system, which had numerous
fundamental design problems which needed solving.  Ren-C is now trying
to go and attack those issues in the same spirit of being able to
express module-like intentions with usermode constructs, that can be
arbitrarily customized.

Modules are loaded only once, so they are not supposed to have
side-effects.  All modules are "isolated", so in order to see
definitions they must be exported...and imported explicitly between
the files.

Some basic binding mechanics (under the umbrella term "Sea of Words")
are now in effect that can help realize modules.  Also, some of the
chaos, redundancy, and uncertainty in the implementation has been
reined in.  However, there are still issues with when imports are
taken from a module, then if those words are imported into the global
scope of a module (as opposed to being accessed through a reference
to the module itself).  These imports are copies of the variable, and
will not see changes...only changes to what is pointed to (which may
change in the referred to module).