Test of using === to write verbose messages
With the "Sea of Words" model, it is no longer the case that every
word used in a file creates a variable.  Previously this was done in
order to make sure that variables named early in the file could be
bound in a way that they'd be linked up to an assignment later in
the file.  That is no longer needed.

https://forum.rebol.info/t/the-sea-of-words/1564

This makes the pleasing-looking section divider something that can
be used with plain words, tuples, and paths.  The existence of forms
ending or beginning in . makes most English text viable:

    === You can write (like this).  And it should work! ===

Previously, out of concern for creating tons of stray variables, one
would have had to have written it in a string:

    === {It might seem (minor) but really, little things add up!} ===

That's still legal to do for those who want it.

In any case, the concept was that these comments that stay resident in
the program code could actually variadically print out the message
as it went by at the higher verbosity levels.  It's an idea that
hasn't really been tried out before, but whitespace seems like a
good place to start.