Add import of Redbol module to the old script
Long term, the goal is to automatically sense from the header whether
emulation is required and pull the appropriate module to configure the
context with compatible emulations of the natives and mezzanine.

For now, just import the redbol emulation layer module explicitly.
(This precludes being able to actually run it in a R3-Alpha/Rebol2/Red)