VERIFY [NULL? ...] not ENSURE NULL
ENSURE NULL as a shorthand for "make sure this thing is not null" was
decided to be too error prone, e.g. if you said "ENSURE TYPE" but had
not assigned TYPE and it was NULL, it would give the wrong answer.

(This is based on a trend to not accept NULL in non-conditional slots.)